Allow appending version information to SLURM_RELEASE
For a long time configure has modified the SLURM Release number
as set in META by stripping off everything before the last '.'
when building the SLURM_VERSION_STRING. This was done so that a
release number of 0.pre1 would become just 'pre1' in the version
string printed by SLURM commands. (e.g. slurm-2.3.0-0.pre1 becomes
slurm-2.3.0-pre1 in sinfo --version).

In attempting to create a new version 2.3.0-2.x of SLURM (branched
from 2.3.0-2), it was found that this method is overzealous, and
results in a version string of just "2.3.0-1" instead of the expected
"2.3.0-2.1". Since the intent of the sed command is only to remove
'0.' from prereleases, this patch makes that explicit, so that
non-prerelease versions branched of tagged SLURM releases keep the
original Release number in the version string.