Revert 35075c10995f4e83d0104662f147cd7b413d25f4
This reverts https://github.com/hautreux/slurm/commit/35075c10995f4e83d0104662f147cd7b413d25f4
My version of RPM doesn't seem to understand the "replace"
parameter for %config, so rpm builds bomb out with
    Invalid %config token: replace
    Invalid %config token: replace

I think the default for %config is to not replace config
files, so I'm not sure this commit is needed. Since the
release agents are not really config files, but need to
be tightly coupled to the release of task/cgroup and
proctrack/cgroup, I would say the release agents should
be moved under /usr/libexec/slurm

At the very least, this commit needs to be reverted before slurm-2.3.1
in order to allow rpms to be built.

Thanks,
Mark Grondona