task/affinity: add multiple_slurmd support in cpuset management
When cpuset is used as the task/affinity backend, a common root
directory is used to group every cpuset subdirectories created
by the plugin. With multiple slurmd support, this result in all
the virtual nodes executing on a same physical node sharing the
same cpuset directories. This patch make sure that every virtual
node has its own root directory in the cpuset hierarchy.