slurmd epilog: add SLURMD_NODENAME in slurmd prolog/epilog env
When multiple slurmd is enabled, there is no way in the prolog
and epilog scripts associated to the slurmd daemons to determine
which node, either virtual or physical is associated to the
execution. SLURMD_NODENAME will provide that information, that is
to say the name of the virtual or physical SLURM node at the origin
of the prolog/epilog execution.