Implement the "--name" option for sacct/accounting.
sacct --name follows the same logic as scancel/squeue, limiting the output only
to jobs having one of the specified name/s.

The job names are simply sent over the wire to be handled by the backend.
Both mysql and filetxt backends were updated to handle this new condition.

This requires an incompatible protocol change in the job_cond structure.
The perl api was also updated.