Permit pending job to exeeded partition limit with QOS flag change.
One of our testers discovered a regression in version 2.3.1.  If a job is
pending due to PartitionNodeLimit and the limit is relieved with a
'sacctmgr modify qos name=<qos name> set flags=partitionmaxnodes' new jobs
exceeding the partition limit (but not the QOS limit) are allowed to run.
However, the pending job is never allowed to run.  Attached is a patch to
address this problem.  FYI, this problem doesn't exist in version 2.4.
Patch from Bill Brophy, Bull.