Major restructure of salloc logic with Loadleveler
This is a major restructuring of the salloc code for use with Loadleveler.
Most of the original code is in the latest srun/load_leveler.c code and
the new salloc logic will be much more light weight and go into src/common
for use from both salloc and srun. It still needs a lot of work, but this
is a new outline of the code.