Rearrange job manager's content to better handle secured FS
Access to secured FS often requires to have a valid token in the
user context. With SLURM, this token can be obtained using one of
the possible pluggable architecture, SPANK or PAM.

IO setup of SLURM can require to access secured FS (stdout/stderr
files). This patch ensures that pluggable frameworks are activated
and called prior to IO setup and that IO are terminated before
calling pluggable framework exit calls.