spank: add spank_option_getopt to spank api
Add a new call to process spank options from a plugin.

The spank_option_getopt() function will search the current
spank environment for use of the option passed as an argument.
The current option cache, and the local environment are checked
for the use of the given spank option. This call is an alternative
to use of a global variable in combination with the option callback,
and is also needed for processing options in the isolated contexts
of slurm_spank_job_prolog() and slurm_spank_job_epilog().