last one
Moe,

After doing more extensive testing, I came to realize
that we had made a bad basic assumption. We believed
that the user's environment should only be what was
sent in the file via the --export-file option.

However, that broke the previous behavior, especially
in regard to Moab jobs. It also caused the SLURM
defined environment variables to be lost as well.

This patch will enable the correct behavior for Moab
on top of SLURM whne using the --export-file option,
but the behavior is less that pefect for using it
stand alone with sbatch. When using the option with
sbatch as a user, the file environment is read in,
and then when the env_array_merge is made, some
variables may get overwritten. This is good for
the SLURM and MPI vairables, but not so good for
others., The problem is trying to reconcile two
sources of environment is very problematic.

I also added a caveat in the man page.

I made changes in my branch of SchedMD SLURM
for 2.3, here is the patch.

Phil