here it is
Moe,

Here it is, I have added a subroutine to env.c to
unset the user's environment and then called it
from sbatch in main. I also removed the comment
from the sbatch man page indicating that it
wasn't working the same for a regular user as
it did for Moab. It should now be functionally
the same.

I think there is still a difference between how
sbatch functions with an environment in a file
than it does from when Moab excve'd the environment.
However, I'm not sure what it would be at this
point.

Again,  for so many iterations....

Phil