2.4 patch - allow slurm user access to a root-only partiiton
Moe,

I have been looking at getting Moab to
run as a non-root user, and it appears
that since the majority of the communicaiton
with slurm is done via the wiki, the only
thing that was not completely working
was allowing a non-root user (when set
as the slurm user) to submit a job to
a "RootOnly=yes" parititon when speicifying
the uid, gid, jobid etc.

The following patch will allow this, and
if you believe it is reasonable, please
put in the 2.4  code base. No hurry on this
one at all.

Phil