task/cgroup: use xcgroup_move_process to move slurmstepd to root memcg
In task_cgroup_memory_fini() the implementation attempts to move
the existing slurmstepd task to the root memory cgroup by writing
the result of getpid(2) to the root memory's 'task' file. This
does not work, however, because slurmstepd is multi-threaded and
thus only the main thread is moved.

This patch replaces the explicit write to 'tasks' with a call to
the new xcgroup_move_process() call, which handles moving all
threads in the process.