Append "*" to default partition name with format and no size
Show this HTML in a new window?
There is a minor problem with the display of partition names in
"sinfo".  Without options, the partition name field displays a
asterisk "*" at the end of the name of the Default partition.  If you
specify a formatting option which contains the %P field specifier with
a width option (e.g., sinfo -o %8P) the asterisk also is appended to
the default partition name.  With no width option, the "%P" displays
the name based on the full length of the name string, however, no "*"
is appended on the default partition name.

The attached patch for version 2.4.0-pre4 corrects the problem so that
the "*" is correctly appended when %P with no width specifier is
used. The patch will also apply to version 2.3.4.

  -Don Albert-