Clarify enforcement of job's --mem option:
I'd like to add the following disclaimer to the documentation of the --mem option to the salloc/sbatch/srun commands. There is currently similar wording in the slurm.conf file, but I've received a bug report in which the memory limits were exceeded (until the next accounting poll).

NOTE: Enforcement of memory limits currently requires enabling of accounting,
which samples memory use on a periodic basis (data need not be stored,  just  collected).
A task may exceed the memory limit until the next periodic accounting sample.

Rod Schultz, Bull