Permit held job to be modified
I was doing some checking to find out why the the 2.4 branch and
master branch of schedmd was not allowing held jobs to be modified,
when attempting to do so,  scontrol would return:

slurm_update error: Requested partition configuration not available now

I did some debugging and found that it was caused by code added to the tail end
of job_limits_check() in job_mgr.c. It  had this addition:

        } else if (job_ptr->priority == 0) {   /* user or administrator hold */
                fail_reason = WAIT_HELD;
        }

It is causes all modifications done by scontrol on held jobs, to fail.