select/cons_res: fix bug in exclusive node allocation
This corrects a bug introduced in commit 952401ec2a56da2b005e6870a0121ccb391736bc
The NODE_CR_RESV value is being stored in a job as an 8-bit variable, so we
can not use a value of 10000 for an exclusive node allocation. It's value was
reduced to 250, which will be the maximum number of jobs that can be run on any
node. Problem demonstrated by "salloc -N1 --exclusive" allocating the job only
one CPU on the node rather than all CPUs on the node.