jobacct_gather plugins: ensure that tot_rss is always consistent with current usage
tot_rss of a step task is computed as the max value of collected RSS for it and its
children. However, tot_rss is the parameter used to aggregate and evaluate the
memory consumption of a set of steps belonging to the same job on a node.
As a result, if multiple steps are present at a time, for example a batch script
and a parallel steps, the limit enforcement of the whole is based on the comparison
between the sum of max usage of each steps and the limit.
It is not correct as the limit should only took into account current usage.

An example of a batch job exhibiting the issue is something like :

 ---
 #!/bin/bash
 #SBATCH -n 8
 #SBATCH --mem-per-cpu 2100

 # sequential action
 # allocate 10 Go and wait 60s to be sure that the tot_rss
 # will be collected (acctgather freq set to 30 in conf)
 ~/malloc -s $((1024*1024)) -n 10 -D 60

 # parallel action
 # use 2 Go per task
 # The job will be killed if there is an error in the global memory
 # consumption evaluation
 srun ~/malloc -s $((1024*1024)) -n 2 -D 1200
 ---

With the patch, the same script will work as expected.